Refactor command!
Depending on the mode of the game, commands (now consistently called actions) need to be aware of the state.current_player_id, which means they need to be handled by the GameServer, not the channel. I’m also trying to avoid creating new functions in more than one place every time there’s a new action.