Randomly distribute all tiles to players
Instead of just one tile each. I’m hoping this will reduce the influence that a lucky first-tile placement (at edges, far from enemies) may have the winner of the game. Like in Risk, now, all players should have somewhat of a choice of where to focus, with greater competition for the better spots.

Naturally, this first commit will need to change, to distribute tiles evenly and not randomly, but this first commit is just a proof-of-concept to try out.