Create domain directory with subdirectories and barebones `adocs`
The domain directory is divided into four parts -- Entities, Collaborators, Actions, and Workflows -- as described in the domain directory's `README.adoc`.

We're numbering the domain subdirectories so that they show up in the desired order. It's not a very appealing way to do this but as of this commit I'm not aware of another option.

We're introducing Asciidoctor with this commit because:

1) From what little I've used it I think it's way cooler (more versatile) than Markdown.

2) This is a highly experimental branch that may never be merged

3) Though I certainly don't want to detract from the main push of this branch with a shiny object, I don't believe having Asciidoctor in here will do much other than illustrate some of the neat things it can do.

Our inclusion of the domain subdirectory `adocs` here may not ultimately be practical -- it's as much showing off Asciidoctor as anything else