Add symlinks for Boards and Regions, start Thinking
Regions seem at first glance hard-pressed to be a first-class entity based on their depth of representation, but instinctively we know they have to be. It is impossible to imagine a group of even mildly experienced Risk players who aren't thinking about regions as they play.

By collecting its representations in one place, Boards show themselves to be problematic. There is indeed a board in Risk -- it's where the tiles and units are. This is what is meant by <div class="Board"> in `board_builder_live`. But it is not what is meant by the "Board" struct or module -- there "Board" is used to refer to what is rightly presented to human players as a "Map" in the new game template.

Map -- GameMap if need be -- makes a great deal more sense. This leaves "Board" free to take its rightful place (if desired) as a container for tiles and selected tile in the game state.

Abstracting a board or map partial out would allow us to remove the new game template from Boards.

Also adds a `THINKING` adoc for some planning. This info will end up in docs and/or commit/PR messages.