Reorganize domain directory
*Nest other domain concepts under Entities*

After some thought we've decided to nest Collaborators, Actions, and Workflows under the Entities they support or whose behavior they describe.

This is being done:

* To improve organization and therefore navigation for Actions and Workflows

* To fix the unsightly domain-folder-name-preceded-by-number issue, which may full well be resolvable by other means

We have added more detail to the first motivation outlined here in the `THINKING`.

*Move and rename general docs on the four `domain` concepts*

These are useful for understanding the purpose of the domain directory. They have been renamed for what they are. As a navigational convenience when viewing on GitHub, `entities.adoc` has been symlinked as `README.adoc` in the `Entities` directory. With the current contents of the other three domain concept base docs and the proposed nested structure of the Entities directory, it will not be optimal to similarly link to these in the nested Collaborator/Action/Workflow directories.

*Make information on four key domain concepts expandable, initializes as collapsed*

This is as much showing off Asciidoctor's functionality as it is truly useful. In favor of this, a person is first presented with the four key concepts and then can drill in (or not) as they like. A person who actively engages with a concept is much more likely to improve their understanding of it than one who is merely presented with it.

In making this change, I found it necessary to remove the title from the included docs and to add it above the expandible example blocks.

The reasoning was twofold:

1) The collapsible example blocks were otherwise showing up with the heading "Details", which is more than somewhat less than ideal.

2) Once expanded, the headings in the included example blocks were showing up raw (with hashtags) rather than being rendered as headings.

I also found it necessary to tag the body contents of each of the included domain concept docs in order to reference said tag to pull in only the body -- referencing section IDs appears only to be supported by interlinks and not include statements.

An unfortunate drawback of making each key domain concept section collapsible is that we have lost the larger headings for each section. I was not able to find a workaround for now.

*The future*

I did foresee a potential future in which an expandable listing of Entities, Collaborators, Actions, and Workflows might live beneath each title, but this idea appears to hit some snags with the proposed nested-under-Entity organization.

A bit much for a commit but we're in serious flux here.